refactor(ide): extract IDE features into separate modules
Extract individual IDE features from lib.rs into dedicated modules:
- completion.rs: Auto-completion logic for fields, fragments, inline fragments
- diagnostics.rs: File and project-wide diagnostics (validation + lint)
- document_symbols.rs: Document symbols and workspace symbol search
- find_references.rs: Find all references for types, fields, fragments
- goto_definition.rs: Go to definition for types, fields, fragments, etc.
- hover.rs: Hover information for symbols

This is the second phase of the graphql-ide modularization (following #377).
Each feature module is now self-contained with its own imports and clear
boundaries, making the codebase easier to navigate and maintain.

Benefits:
- Each feature has clear ownership in its own module
- Analysis methods now delegate to focused module functions
- Easier to test individual features in isolation
- Better code organization following rust-analyzer patterns
